requirements.txt
customtkinter
Pillow
python-dotenv
requests
terragen-rpc
scipy
opensimplex
